feat: refactor codes and read configurations from pydantic
