--- llama_deploy/message_queues/solace/__init__.py
@@ -1,3 +1,0 @@
-from .solace import SolaceMessageQueue, SolaceMessageQueueConfig
-
-__all__ = ["SolaceMessageQueue", "SolaceMessageQueueConfig"]
